basic pawn movement (has bugs) and piece capture

+++ src/Bishop.cpp
@@ -1,6 +1,6 @@
 #include "Bishop.h"
 
-Bishop::Bishop(SDL_Renderer* m_Renderer, bool m_pieceTeam, int m_XPos, int m_YPos)
+Bishop::Bishop(SDL_Renderer* m_Renderer, bool m_pieceTeam, float m_XPos, float m_YPos)
 	: Piece(m_Renderer, BISHOP, m_pieceTeam, m_XPos, m_YPos)
 {
