+++ src/Rook.cpp
@@ -1,6 +1,6 @@
 #include "Rook.h"
 
-Rook::Rook(SDL_Renderer* m_Renderer, bool m_pieceTeam, int m_XPos, int m_YPos)
+Rook::Rook(SDL_Renderer* m_Renderer, bool m_pieceTeam, float m_XPos, float m_YPos)
 	: Piece(m_Renderer, ROOK, m_pieceTeam, m_XPos, m_YPos)
 {
